+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated citeproc handling for pandoc >=2.11.  Default is now to use the
+  `--citeproc` flag on the `pandoc` command line internally.
 
 ## [0.6.3] - 2020-05-27
 ### Fixed